feat: move fetch
