+++ packages/discord.js/src/index.js
@@ -56,7 +56,7 @@ exports.ChannelManager = require('./managers/ChannelManager');
 exports.ClientVoiceManager = require('./client/voice/ClientVoiceManager');
 exports.DataManager = require('./managers/DataManager');
 exports.GuildApplicationCommandManager = require('./managers/GuildApplicationCommandManager');
-exports.GuildAuditLogEntryManager = require('./managers/GuildAuditLogEntryManager');
+exports.GuildAuditLogManager = require('./managers/GuildAuditLogManager');
 exports.GuildBanManager = require('./managers/GuildBanManager');
 exports.GuildChannelManager = require('./managers/GuildChannelManager');
 exports.GuildEmojiManager = require('./managers/GuildEmojiManager');
